🤡♻️🔒️ Refactor, secure, and remove mocks on profile
