+++ src/mocks/client/handlers.ts
@@ -1,5 +1,4 @@
 import { authHandlers } from './auth.handlers';
-import { userHandlers } from './users.handlers';
 
 // Mocks for client side requests.
-export const handlers = [...authHandlers, ...userHandlers];
+export const handlers = [...authHandlers];
